Clean up gui-cli
