+++ jabgui/src/main/java/org/jabref/cli/ArgumentProcessor.java
@@ -1,75 +1,31 @@
 package org.jabref.cli;
 
 import java.io.IOException;
-import java.io.OutputStreamWriter;
-import java.io.Writer;
 import java.net.MalformedURLException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import java.util.Optional;
-import java.util.Set;
 import java.util.prefs.BackingStoreException;
 
-import org.jabref.logic.FilePreferences;
-import org.jabref.logic.JabRefException;
+import org.jabref.gui.preferences.GuiPreferences;
 import org.jabref.logic.UiCommand;
-import org.jabref.logic.bibtex.FieldPreferences;
-import org.jabref.logic.citationkeypattern.CitationKeyGenerator;
-import org.jabref.logic.exporter.AtomicFileWriter;
-import org.jabref.logic.exporter.BibDatabaseWriter;
-import org.jabref.logic.exporter.BibWriter;
-import org.jabref.logic.exporter.BibtexDatabaseWriter;
-import org.jabref.logic.exporter.EmbeddedBibFilePdfExporter;
-import org.jabref.logic.exporter.Exporter;
-import org.jabref.logic.exporter.ExporterFactory;
-import org.jabref.logic.exporter.SelfContainedSaveConfiguration;
-import org.jabref.logic.exporter.XmpPdfExporter;
 import org.jabref.logic.importer.FetcherException;
 import org.jabref.logic.importer.ImportException;
-import org.jabref.logic.importer.ImportFormatPreferences;
 import org.jabref.logic.importer.ImportFormatReader;
 import org.jabref.logic.importer.OpenDatabase;
-import org.jabref.logic.importer.ParseException;
 import org.jabref.logic.importer.ParserResult;
-import org.jabref.logic.importer.SearchBasedFetcher;
-import org.jabref.logic.importer.WebFetchers;
-import org.jabref.logic.importer.fileformat.BibtexParser;
-import org.jabref.logic.journals.JournalAbbreviationRepository;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.net.URLDownload;
 import org.jabref.logic.os.OS;
-import org.jabref.logic.preferences.CliPreferences;
-import org.jabref.logic.quality.consistency.BibliographyConsistencyCheck;
-import org.jabref.logic.quality.consistency.BibliographyConsistencyCheckResultCsvWriter;
-import org.jabref.logic.quality.consistency.BibliographyConsistencyCheckResultTxtWriter;
-import org.jabref.logic.quality.consistency.BibliographyConsistencyCheckResultWriter;
-import org.jabref.logic.search.DatabaseSearcher;
-import org.jabref.logic.search.PostgreServer;
-import org.jabref.logic.search.SearchPreferences;
 import org.jabref.logic.shared.prefs.SharedDatabasePreferences;
-import org.jabref.logic.util.CurrentThreadTaskExecutor;
-import org.jabref.logic.util.io.FileUtil;
-import org.jabref.logic.xmp.XmpPreferences;
-import org.jabref.model.database.BibDatabase;
-import org.jabref.model.database.BibDatabaseContext;
-import org.jabref.model.database.BibDatabaseMode;
-import org.jabref.model.entry.BibEntry;
-import org.jabref.model.entry.BibEntryTypesManager;
-import org.jabref.model.search.query.SearchQuery;
-import org.jabref.model.strings.StringUtil;
 import org.jabref.model.util.DummyFileUpdateMonitor;
 import org.jabref.model.util.FileUpdateMonitor;
 
-import com.airhacks.afterburner.injection.Injector;
-import com.google.common.base.Throwables;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-///  TODO: This is a clone of {@link org.jabref.cli.ArgumentProcessor} in jabkit - should be unified
 public class ArgumentProcessor {
     private static final Logger LOGGER = LoggerFactory.getLogger(ArgumentProcessor.class);
 
@@ -79,9 +35,8 @@ public class ArgumentProcessor {
 
     private final Mode startupMode;
 
-    private final CliPreferences cliPreferences;
+    private final GuiPreferences preferences;
     private final FileUpdateMonitor fileUpdateMonitor;
-    private final BibEntryTypesManager entryTypesManager;
 
     private boolean guiNeeded;
     private final List<UiCommand> uiCommands = new ArrayList<>();
@@ -92,15 +47,13 @@ public class ArgumentProcessor {
      */
     public ArgumentProcessor(String[] args,
                              Mode startupMode,
-                             CliPreferences cliPreferences,
-                             FileUpdateMonitor fileUpdateMonitor,
-                             BibEntryTypesManager entryTypesManager)
+                             GuiPreferences preferences,
+                             FileUpdateMonitor fileUpdateMonitor)
             throws org.apache.commons.cli.ParseException {
         this.cli = new CliOptions(args);
         this.startupMode = startupMode;
-        this.cliPreferences = cliPreferences;
+        this.preferences = preferences;
         this.fileUpdateMonitor = fileUpdateMonitor;
-        this.entryTypesManager = entryTypesManager;
     }
 
     /**
@@ -112,19 +65,6 @@ public class ArgumentProcessor {
         Optional<ParserResult> result = importFile(importArguments);
         result.ifPresent(ParserResult::setToOpenTab);
         return result;
-    }
-
-    private Optional<ParserResult> importBibtexToOpenBase(String argument, ImportFormatPreferences importFormatPreferences) {
-        BibtexParser parser = new BibtexParser(importFormatPreferences);
-        try {
-            List<BibEntry> entries = parser.parseEntries(argument);
-            ParserResult result = new ParserResult(entries);
-            result.setToOpenTab();
-            return Optional.of(result);
-        } catch (ParseException e) {
-            System.err.println(Localization.lang("Error occurred when parsing entry") + ": " + e.getLocalizedMessage());
-            return Optional.empty();
-        }
     }
 
     /**
@@ -172,9 +112,9 @@ public class ArgumentProcessor {
     private Optional<ParserResult> importFile(Path file, String importFormat) {
         try {
             ImportFormatReader importFormatReader = new ImportFormatReader(
-                    cliPreferences.getImporterPreferences(),
-                    cliPreferences.getImportFormatPreferences(),
-                    cliPreferences.getCitationKeyPatternPreferences(),
+                    preferences.getImporterPreferences(),
+                    preferences.getImportFormatPreferences(),
+                    preferences.getCitationKeyPatternPreferences(),
                     fileUpdateMonitor
             );
 
@@ -200,88 +140,26 @@ public class ArgumentProcessor {
 
     public void processArguments() {
         uiCommands.clear();
+        guiNeeded = true;
 
         if ((startupMode == Mode.INITIAL_START) && cli.isShowVersion()) {
             cli.displayVersion();
-        }
-
-        if ((startupMode == Mode.INITIAL_START) && cli.isHelp()) {
-            CliOptions.printUsage(cliPreferences);
             guiNeeded = false;
             return;
         }
 
-        guiNeeded = true;
+        if ((startupMode == Mode.INITIAL_START) && cli.isHelp()) {
+            CliOptions.printUsage(preferences);
+            guiNeeded = false;
+            return;
+        }
 
         // Check if we should reset all preferences to default values:
         if (cli.isPreferencesReset()) {
-            resetPreferences(cli.getPreferencesReset());
-        }
-
-        // Check if we should import preferences from a file:
-        if (cli.isPreferencesImport()) {
-            importPreferences();
+            resetPreferences();
         }
 
         List<ParserResult> loaded = importAndOpenFiles();
-
-        if (!cli.isBlank() && cli.isFetcherEngine()) {
-            fetch(cli.getFetcherEngine()).ifPresent(loaded::add);
-        }
-
-        if (cli.isExportMatches()) {
-            if (!loaded.isEmpty()) {
-                if (!exportMatches(loaded)) {
-                    return;
-                }
-            } else {
-                System.err.println(Localization.lang("The output option depends on a valid input option."));
-            }
-        }
-
-        if (cli.isGenerateCitationKeys()) {
-            regenerateCitationKeys(loaded);
-        }
-
-        if ((cli.isWriteXmpToPdf() && cli.isEmbedBibFileInPdf()) || (cli.isWriteMetadataToPdf() && (cli.isWriteXmpToPdf() || cli.isEmbedBibFileInPdf()))) {
-            System.err.println("Give only one of [writeXmpToPdf, embedBibFileInPdf, writeMetadataToPdf]");
-        }
-
-        if (cli.isWriteMetadataToPdf() || cli.isWriteXmpToPdf() || cli.isEmbedBibFileInPdf()) {
-            if (!loaded.isEmpty()) {
-                writeMetadataToPdf(loaded,
-                        cli.getWriteMetadataToPdf(),
-                        cliPreferences.getXmpPreferences(),
-                        cliPreferences.getFilePreferences(),
-                        cliPreferences.getLibraryPreferences().getDefaultBibDatabaseMode(),
-                        cliPreferences.getCustomEntryTypesRepository(),
-                        cliPreferences.getFieldPreferences(),
-                        Injector.instantiateModelOrService(JournalAbbreviationRepository.class),
-                        cli.isWriteXmpToPdf() || cli.isWriteMetadataToPdf(),
-                        cli.isEmbedBibFileInPdf() || cli.isWriteMetadataToPdf());
-            }
-        }
-
-        if (cli.isFileExport()) {
-            if (!loaded.isEmpty()) {
-                exportFile(loaded, cli.getFileExport().split(","));
-                LOGGER.debug("Finished export");
-            } else {
-                System.err.println(Localization.lang("The output option depends on a valid import option."));
-            }
-        }
-
-        if (cli.isPreferencesExport()) {
-            try {
-                cliPreferences.exportPreferences(Path.of(cli.getPreferencesExport()));
-            } catch (JabRefException ex) {
-                LOGGER.error("Cannot export preferences", ex);
-            }
-        }
-
-        if (!cli.isBlank() && cli.isAuxImport()) {
-            doAuxImport(loaded);
-        }
 
         if (cli.isBlank()) {
             uiCommands.add(new UiCommand.BlankWorkspace());
@@ -297,305 +175,6 @@ public class ArgumentProcessor {
 
         if (cli.isBlank() && loaded.isEmpty()) {
             uiCommands.add(new UiCommand.BlankWorkspace());
-        }
-
-        if (cli.isCheckConsistency()) {
-            checkConsistency(cliPreferences, entryTypesManager);
-        }
-    }
-
-    private void checkConsistency(CliPreferences cliPreferences,
-                                  BibEntryTypesManager entryTypesManager) {
-        Optional<String> fileName = Optional.ofNullable(cli.getCheckConsistency());
-
-        if (fileName.isEmpty()) {
-            System.out.println(Localization.lang("No file specified for consistency check."));
-            return;
-        }
-
-        Optional<String> outputFormat = Optional.ofNullable(cli.getCheckConsistencyOutputFormat());
-
-        Path filePath = Path.of(fileName.get());
-        ParserResult pr;
-        try {
-            pr = OpenDatabase.loadDatabase(filePath, cliPreferences.getImportFormatPreferences(), fileUpdateMonitor);
-        } catch (IOException ex) {
-            LOGGER.error("Error reading '{}'.", filePath, ex);
-            return;
-        }
-        BibDatabaseContext databaseContext = pr.getDatabaseContext();
-        List<BibEntry> entries = databaseContext.getDatabase().getEntries();
-
-        BibliographyConsistencyCheck consistencyCheck = new BibliographyConsistencyCheck();
-        BibliographyConsistencyCheck.Result result = consistencyCheck.check(entries);
-
-        Writer writer = new OutputStreamWriter(System.out);
-        BibliographyConsistencyCheckResultWriter checkResultWriter;
-        if (outputFormat.isEmpty() || "txt".equalsIgnoreCase(outputFormat.get())) {
-            checkResultWriter = new BibliographyConsistencyCheckResultTxtWriter(
-                    result,
-                    writer,
-                    cli.isPorcelainOutputMode(),
-                    entryTypesManager,
-                    databaseContext.getMode());
-        } else {
-            checkResultWriter = new BibliographyConsistencyCheckResultCsvWriter(
-                    result,
-                    writer,
-                    cli.isPorcelainOutputMode(),
-                    entryTypesManager,
-                    databaseContext.getMode());
-        }
-
-        // System.out should not be closed, therefore no try-with-resources
-        try {
-            checkResultWriter.writeFindings();
-            writer.flush();
-        } catch (IOException e) {
-            LOGGER.error("Error writing results", e);
-        }
-        if (!cli.isPorcelainOutputMode()) {
-            System.out.println(Localization.lang("Consistency check completed"));
-        }
-    }
-
-    private static void writeMetadataToPdf(List<ParserResult> loaded,
-                                    String filesAndCiteKeys,
-                                    XmpPreferences xmpPreferences,
-                                    FilePreferences filePreferences,
-                                    BibDatabaseMode databaseMode,
-                                    BibEntryTypesManager entryTypesManager,
-                                    FieldPreferences fieldPreferences,
-                                    JournalAbbreviationRepository abbreviationRepository,
-                                    boolean writeXMP,
-                                    boolean embeddBibfile) {
-        if (loaded.isEmpty()) {
-            LOGGER.error("The write xmp option depends on a valid import option.");
-            return;
-        }
-        ParserResult pr = loaded.getLast();
-        BibDatabaseContext databaseContext = pr.getDatabaseContext();
-
-        XmpPdfExporter xmpPdfExporter = new XmpPdfExporter(xmpPreferences);
-        EmbeddedBibFilePdfExporter embeddedBibFilePdfExporter = new EmbeddedBibFilePdfExporter(databaseMode, entryTypesManager, fieldPreferences);
-
-        if ("all".equals(filesAndCiteKeys)) {
-            for (BibEntry entry : databaseContext.getEntries()) {
-                writeMetadataToPDFsOfEntry(
-                        databaseContext,
-                        entry.getCitationKey().orElse("<no cite key defined>"),
-                        entry,
-                        filePreferences,
-                        xmpPdfExporter,
-                        embeddedBibFilePdfExporter,
-                        abbreviationRepository,
-                        writeXMP,
-                        embeddBibfile);
-            }
-            return;
-        }
-
-        List<String> citeKeys = new ArrayList<>();
-        List<String> pdfs = new ArrayList<>();
-        for (String fileOrCiteKey : filesAndCiteKeys.split(",")) {
-            if (fileOrCiteKey.toLowerCase(Locale.ROOT).endsWith(".pdf")) {
-                pdfs.add(fileOrCiteKey);
-            } else {
-                citeKeys.add(fileOrCiteKey);
-            }
-        }
-
-        writeMetadataToPdfByCitekey(
-                databaseContext,
-                citeKeys,
-                filePreferences,
-                xmpPdfExporter,
-                embeddedBibFilePdfExporter,
-                abbreviationRepository,
-                writeXMP,
-                embeddBibfile);
-        writeMetadataToPdfByFileNames(
-                databaseContext,
-                pdfs,
-                filePreferences,
-                xmpPdfExporter,
-                embeddedBibFilePdfExporter,
-                abbreviationRepository,
-                writeXMP,
-                embeddBibfile);
-    }
-
-    private static void writeMetadataToPDFsOfEntry(BibDatabaseContext databaseContext,
-                                            String citeKey,
-                                            BibEntry entry,
-                                            FilePreferences filePreferences,
-                                            XmpPdfExporter xmpPdfExporter,
-                                            EmbeddedBibFilePdfExporter embeddedBibFilePdfExporter,
-                                            JournalAbbreviationRepository abbreviationRepository,
-                                            boolean writeXMP,
-                                            boolean embedBibfile) {
-        try {
-            if (writeXMP) {
-                if (xmpPdfExporter.exportToAllFilesOfEntry(databaseContext, filePreferences, entry, List.of(entry), abbreviationRepository)) {
-                    System.out.printf("Successfully written XMP metadata on at least one linked file of %s%n", citeKey);
-                } else {
-                    System.err.printf("Cannot write XMP metadata on any linked files of %s. Make sure there is at least one linked file and the path is correct.%n", citeKey);
-                }
-            }
-            if (embedBibfile) {
-                if (embeddedBibFilePdfExporter.exportToAllFilesOfEntry(databaseContext, filePreferences, entry, List.of(entry), abbreviationRepository)) {
-                    System.out.printf("Successfully embedded metadata on at least one linked file of %s%n", citeKey);
-                } else {
-                    System.out.printf("Cannot embed metadata on any linked files of %s. Make sure there is at least one linked file and the path is correct.%n", citeKey);
-                }
-            }
-        } catch (Exception e) {
-            LOGGER.error("Failed writing metadata on a linked file of {}.", citeKey);
-        }
-    }
-
-    private static void writeMetadataToPdfByCitekey(BibDatabaseContext databaseContext,
-                                             List<String> citeKeys,
-                                             FilePreferences filePreferences,
-                                             XmpPdfExporter xmpPdfExporter,
-                                             EmbeddedBibFilePdfExporter embeddedBibFilePdfExporter,
-                                             JournalAbbreviationRepository abbreviationRepository,
-                                             boolean writeXMP,
-                                             boolean embeddBibfile) {
-        for (String citeKey : citeKeys) {
-            List<BibEntry> bibEntryList = databaseContext.getDatabase().getEntriesByCitationKey(citeKey);
-            if (bibEntryList.isEmpty()) {
-                System.err.printf("Skipped - Cannot find %s in library.%n", citeKey);
-                continue;
-            }
-            for (BibEntry entry : bibEntryList) {
-                writeMetadataToPDFsOfEntry(databaseContext, citeKey, entry, filePreferences, xmpPdfExporter, embeddedBibFilePdfExporter, abbreviationRepository, writeXMP, embeddBibfile);
-            }
-        }
-    }
-
-    private static void writeMetadataToPdfByFileNames(BibDatabaseContext databaseContext,
-                                               List<String> pdfs,
-                                               FilePreferences filePreferences,
-                                               XmpPdfExporter xmpPdfExporter,
-                                               EmbeddedBibFilePdfExporter embeddedBibFilePdfExporter,
-                                               JournalAbbreviationRepository abbreviationRepository,
-                                               boolean writeXMP,
-                                               boolean embeddBibfile) {
-        for (String fileName : pdfs) {
-            Path filePath = Path.of(fileName);
-            if (!filePath.isAbsolute()) {
-                filePath = FileUtil.find(fileName, databaseContext.getFileDirectories(filePreferences)).orElse(FileUtil.find(fileName, List.of(Path.of("").toAbsolutePath())).orElse(filePath));
-            }
-            if (Files.exists(filePath)) {
-                try {
-                    if (writeXMP) {
-                        if (xmpPdfExporter.exportToFileByPath(databaseContext, filePreferences, filePath, abbreviationRepository)) {
-                            System.out.printf("Successfully written XMP metadata of at least one entry to %s%n", fileName);
-                        } else {
-                            System.out.printf("File %s is not linked to any entry in database.%n", fileName);
-                        }
-                    }
-                    if (embeddBibfile) {
-                        if (embeddedBibFilePdfExporter.exportToFileByPath(databaseContext, filePreferences, filePath, abbreviationRepository)) {
-                            System.out.printf("Successfully embedded XMP metadata of at least one entry to %s%n", fileName);
-                        } else {
-                            System.out.printf("File %s is not linked to any entry in database.%n", fileName);
-                        }
-                    }
-                } catch (IOException e) {
-                    LOGGER.error("Error accessing file '{}'.", fileName);
-                } catch (Exception e) {
-                    LOGGER.error("Error writing entry to {}.", fileName);
-                }
-            } else {
-                LOGGER.error("Skipped - PDF {} does not exist", fileName);
-            }
-        }
-    }
-
-    private boolean exportMatches(List<ParserResult> loaded) {
-        String[] data = cli.getExportMatches().split(",");
-        String searchTerm = data[0].replace("\\$", " "); // enables blanks within the search term:
-        // $ stands for a blank
-        ParserResult pr = loaded.getLast();
-        BibDatabaseContext databaseContext = pr.getDatabaseContext();
-
-        SearchPreferences searchPreferences = cliPreferences.getSearchPreferences();
-        SearchQuery query = new SearchQuery(searchTerm, searchPreferences.getSearchFlags());
-
-        List<BibEntry> matches;
-        try {
-            // extract current thread task executor from indexManager
-            matches = new DatabaseSearcher(query, databaseContext, new CurrentThreadTaskExecutor(), cliPreferences, Injector.instantiateModelOrService(PostgreServer.class)).getMatches();
-        } catch (IOException e) {
-            LOGGER.error("Error occurred when searching", e);
-            return false;
-        }
-
-        // export matches
-        if (!matches.isEmpty()) {
-            String formatName;
-
-            // read in the export format, take default format if no format entered
-            switch (data.length) {
-                case 3 -> formatName = data[2];
-                case 2 ->
-                        // default exporter: bib file
-                        formatName = "bib";
-                default -> {
-                    System.err.println(Localization.lang("Output file missing").concat(". \n \t ")
-                                                   .concat(Localization.lang("Usage")).concat(": ") + CliOptions.getExportMatchesSyntax());
-                    guiNeeded = false;
-                    return false;
-                }
-            }
-
-            if ("bib".equals(formatName)) {
-                // output a bib file as default or if
-                // provided exportFormat is "bib"
-                saveDatabase(new BibDatabase(matches), data[1]);
-                LOGGER.debug("Finished export");
-            } else {
-                // export new database
-                ExporterFactory exporterFactory = ExporterFactory.create(cliPreferences);
-                Optional<Exporter> exporter = exporterFactory.getExporterByName(formatName);
-                if (exporter.isEmpty()) {
-                    System.err.println(Localization.lang("Unknown export format %0", formatName));
-                } else {
-                    // We have an TemplateExporter instance:
-                    try {
-                        System.out.println(Localization.lang("Exporting %0", data[1]));
-                        exporter.get().export(
-                                databaseContext,
-                                Path.of(data[1]),
-                                matches,
-                                List.of(),
-                                Injector.instantiateModelOrService(JournalAbbreviationRepository.class));
-                    } catch (Exception ex) {
-                        System.err.println(Localization.lang("Could not export file '%0' (reason: %1)", data[1], Throwables.getStackTraceAsString(ex)));
-                    }
-                }
-            }
-        } else {
-            System.err.println(Localization.lang("No search matches."));
-        }
-        return true;
-    }
-
-    private void doAuxImport(List<ParserResult> loaded) {
-        boolean usageMsg;
-
-        if (!loaded.isEmpty()) {
-            usageMsg = generateAux(loaded, cli.getAuxImport().split(","));
-        } else {
-            usageMsg = true;
-        }
-
-        if (usageMsg) {
-            System.out.println(Localization.lang("no base-BibTeX-file specified!"));
-            System.out.println(Localization.lang("usage") + " :");
-            System.out.println("jabref --aux infile[.aux],outfile[.bib] base-BibTeX-file");
         }
     }
 
@@ -617,7 +196,7 @@ public class ArgumentProcessor {
                     try {
                         pr = OpenDatabase.loadDatabase(
                                 Path.of(aLeftOver),
-                                cliPreferences.getImportFormatPreferences(),
+                                preferences.getImportFormatPreferences(),
                                 fileUpdateMonitor);
                         // In contrast to org.jabref.gui.LibraryTab.onDatabaseLoadingSucceed, we do not execute OpenDatabaseAction.performPostOpenActions(result, dialogService);
                     } catch (IOException ex) {
@@ -656,204 +235,22 @@ public class ArgumentProcessor {
             importToOpenBase(cli.getImportToOpenBase()).ifPresent(loaded::add);
         }
 
-        if (!cli.isBlank() && cli.isBibtexImport()) {
-            importBibtexToOpenBase(cli.getBibtexImport(), cliPreferences.getImportFormatPreferences()).ifPresent(loaded::add);
-        }
-
         return loaded;
     }
 
-    private boolean generateAux(List<ParserResult> loaded, String[] data) {
-        if (data.length == 2) {
-            ParserResult pr = loaded.getFirst();
-            AuxCommandLine acl = new AuxCommandLine(data[0], pr.getDatabase());
-            BibDatabase newBase = acl.perform();
-
-            boolean notSavedMsg = false;
-
-            // write an output, if something could be resolved
-            if ((newBase != null) && newBase.hasEntries()) {
-                String subName = StringUtil.getCorrectFileName(data[1], "bib");
-                saveDatabase(newBase, subName);
-                notSavedMsg = true;
-            }
-
-            if (!notSavedMsg) {
-                System.out.println(Localization.lang("no library generated"));
-            }
-            return false;
-        } else {
-            return true;
-        }
-    }
-
-    private void saveDatabase(BibDatabase newBase, String subName) {
+    private void resetPreferences() {
         try {
-            System.out.println(Localization.lang("Saving") + ": " + subName);
-            try (AtomicFileWriter fileWriter = new AtomicFileWriter(Path.of(subName), StandardCharsets.UTF_8)) {
-                BibWriter bibWriter = new BibWriter(fileWriter, OS.NEWLINE);
-                SelfContainedSaveConfiguration saveConfiguration = (SelfContainedSaveConfiguration) new SelfContainedSaveConfiguration()
-                        .withReformatOnSave(cliPreferences.getLibraryPreferences().shouldAlwaysReformatOnSave());
-                BibDatabaseWriter databaseWriter = new BibtexDatabaseWriter(
-                        bibWriter,
-                        saveConfiguration,
-                        cliPreferences.getFieldPreferences(),
-                        cliPreferences.getCitationKeyPatternPreferences(),
-                        entryTypesManager);
-                databaseWriter.saveDatabase(new BibDatabaseContext(newBase));
-
-                // Show just a warning message if encoding did not work for all characters:
-                if (fileWriter.hasEncodingProblems()) {
-                    System.err.println(Localization.lang("Warning") + ": "
-                                       + Localization.lang("UTF-8 could not be used to encode the following characters: %0", fileWriter.getEncodingProblems()));
-                }
-            }
-        } catch (IOException ex) {
-            System.err.println(Localization.lang("Could not save file.") + "\n" + ex.getLocalizedMessage());
-        }
-    }
-
-    private void exportFile(List<ParserResult> loaded, String[] data) {
-        if (data.length == 1) {
-            // This signals that the latest import should be stored in BibTeX
-            // format to the given file.
-            if (!loaded.isEmpty()) {
-                ParserResult pr = loaded.getLast();
-                if (!pr.isInvalid()) {
-                    saveDatabase(pr.getDatabase(), data[0]);
-                }
-            } else {
-                System.err.println(Localization.lang("The output option depends on a valid import option."));
-            }
-        } else if (data.length == 2) {
-            // This signals that the latest import should be stored in the given
-            // format to the given file.
-            ParserResult parserResult = loaded.getLast();
-
-            Path path = parserResult.getPath().get().toAbsolutePath();
-            BibDatabaseContext databaseContext = parserResult.getDatabaseContext();
-            databaseContext.setDatabasePath(path);
-            List<Path> fileDirForDatabase = databaseContext
-                    .getFileDirectories(cliPreferences.getFilePreferences());
-            System.out.println(Localization.lang("Exporting %0", data[0]));
-            ExporterFactory exporterFactory = ExporterFactory.create(cliPreferences);
-            Optional<Exporter> exporter = exporterFactory.getExporterByName(data[1]);
-            if (exporter.isEmpty()) {
-                System.err.println(Localization.lang("Unknown export format %0", data[1]));
-            } else {
-                // We have an exporter:
-                try {
-                    exporter.get().export(
-                            parserResult.getDatabaseContext(),
-                            Path.of(data[0]),
-                            parserResult.getDatabaseContext().getDatabase().getEntries(),
-                            fileDirForDatabase,
-                            Injector.instantiateModelOrService(JournalAbbreviationRepository.class));
-                } catch (Exception ex) {
-                    System.err.println(Localization.lang("Could not export file '%0' (reason: %1)", data[0], Throwables.getStackTraceAsString(ex)));
-                }
-            }
-        }
-    }
-
-    private void importPreferences() {
-        try {
-            cliPreferences.importPreferences(Path.of(cli.getPreferencesImport()));
-            Injector.setModelOrService(BibEntryTypesManager.class, cliPreferences.getCustomEntryTypesRepository());
-        } catch (JabRefException ex) {
-            LOGGER.error("Cannot import preferences", ex);
-        }
-    }
-
-    private void resetPreferences(String value) {
-        if ("all".equals(value.trim())) {
-            try {
-                System.out.println(Localization.lang("Setting all preferences to default values."));
-                cliPreferences.clear();
-                new SharedDatabasePreferences().clear();
-            } catch (BackingStoreException e) {
-                System.err.println(Localization.lang("Unable to clear preferences."));
-                LOGGER.error("Unable to clear preferences", e);
-            }
-        } else {
-            String[] keys = value.split(",");
-            for (String key : keys) {
-                try {
-                    cliPreferences.deleteKey(key.trim());
-                    System.out.println(Localization.lang("Resetting preference key '%0'", key.trim()));
-                } catch (IllegalArgumentException e) {
-                    System.out.println(e.getMessage());
-                }
-            }
-        }
-    }
-
-    private void regenerateCitationKeys(List<ParserResult> loaded) {
-        for (ParserResult parserResult : loaded) {
-            BibDatabase database = parserResult.getDatabase();
-
-            LOGGER.info(Localization.lang("Regenerating citation keys according to metadata"));
-
-            CitationKeyGenerator keyGenerator = new CitationKeyGenerator(
-                    parserResult.getDatabaseContext(),
-                    cliPreferences.getCitationKeyPatternPreferences());
-            for (BibEntry entry : database.getEntries()) {
-                keyGenerator.generateAndSetKey(entry);
-            }
-        }
-    }
-
-    /**
-     * Run an entry fetcher from the command line.
-     *
-     * @param fetchCommand A string containing both the name of the fetcher to use and the search query, separated by a :
-     * @return A parser result containing the entries fetched or null if an error occurred.
-     */
-    private Optional<ParserResult> fetch(String fetchCommand) {
-        if ((fetchCommand == null) || !fetchCommand.contains(":")) {
-            System.out.println(Localization.lang("Expected syntax for --fetch='<name of fetcher>:<query>'"));
-            System.out.println(Localization.lang("The following fetchers are available:"));
-            return Optional.empty();
-        }
-
-        String[] split = fetchCommand.split(":");
-        String engine = split[0];
-        String query = split[1];
-
-        Set<SearchBasedFetcher> fetchers = WebFetchers.getSearchBasedFetchers(
-                cliPreferences.getImportFormatPreferences(),
-                cliPreferences.getImporterPreferences());
-        Optional<SearchBasedFetcher> selectedFetcher = fetchers.stream()
-                                                               .filter(fetcher -> fetcher.getName().equalsIgnoreCase(engine))
-                                                               .findFirst();
-        if (selectedFetcher.isEmpty()) {
-            System.out.println(Localization.lang("Could not find fetcher '%0'", engine));
-
-            System.out.println(Localization.lang("The following fetchers are available:"));
-            fetchers.forEach(fetcher -> System.out.println("  " + fetcher.getName()));
-
-            return Optional.empty();
-        } else {
-            System.out.println(Localization.lang("Running query '%0' with fetcher '%1'.", query, engine));
-            System.out.print(Localization.lang("Please wait..."));
-            try {
-                List<BibEntry> matches = selectedFetcher.get().performSearch(query);
-                if (matches.isEmpty()) {
-                    System.out.println("\r" + Localization.lang("No results found."));
-                    return Optional.empty();
-                } else {
-                    System.out.println("\r" + Localization.lang("Found %0 results.", String.valueOf(matches.size())));
-                    return Optional.of(new ParserResult(matches));
-                }
-            } catch (FetcherException e) {
-                LOGGER.error("Error while fetching", e);
-                return Optional.empty();
-            }
+            System.out.println(Localization.lang("Setting all preferences to default values."));
+            preferences.clear();
+            new SharedDatabasePreferences().clear();
+        } catch (BackingStoreException e) {
+            System.err.println(Localization.lang("Unable to clear preferences."));
+            LOGGER.error("Unable to clear preferences", e);
         }
     }
 
     public boolean shouldShutDown() {
-        return cli.isDisableGui() || cli.isShowVersion() || !guiNeeded;
+        return cli.isShowVersion() || !guiNeeded;
     }
 
     public List<UiCommand> getUiCommands() {
